Implémentation couche DAO
! p.14 : peut être besoin de écrire une requete JPQL pour récupérer un
employé par SS (champ indemnité non initialisé)

+++ server-pam/src/main/java/pam/repositories/CotisationRepository.java
@@ -3,5 +3,5 @@ package pam.repositories;
 import org.springframework.data.repository.CrudRepository;
 import pam.entities.Cotisation;
 
-public interface CotisationRepository extends CrudRepository<Cotisation, Double> {
+public interface CotisationRepository extends CrudRepository<Cotisation, Long> {
 }
