+++ server-pam/src/main/java/pam/repositories/IndemniteRepository.java
@@ -3,5 +3,5 @@ package pam.repositories;
 import org.springframework.data.repository.CrudRepository;
 import pam.entities.Indemnite;
 
-public interface IndemniteRepository extends CrudRepository<Indemnite, Double>{
+public interface IndemniteRepository extends CrudRepository<Indemnite, Long>{
 }
